Add tensor tests to standalone module
